Make Scribe actually an optional dependency
Up until now even if you used a different tracer (Kafka, JSON) you still
needed to depend on Scribe because of the hard coded require at the top
of the rack handler file.
This commit moves the require directives to the appropriate places so
that we require only what we need.