Moves CI to GitHub Actions
This moves from Travis to GitHub Actions based test and deploy workflow.

Currently, this does not include testing on JRuby, but that could be
added in the future.

Notably, this uses standard script conventions used by other projects in
OpenZipkin. The main change to the actual code here is reverting
version.rb to 0.0.0 like some other Ruby projects are doing, instead of
updating it manually in git each time.